Rename Solum<T> into Only<T>
In other languages the Option type has been represented by a pair Just or Nothing and a pair Some or None.

As JavaScript has both null and undefined types, we need to support three pairs. So originally for Nullable the present value was named as `Solum<T>`

It has been an unfortunate choice, as there is no actual word “Solum”. “Only” is a much better choice, as it can be defined as “and no one or nothing more besides; solely or exclusively,” which is quite similar to “Just.”

Switching to `Only<T>` returns code to the goal of having English-readable text.