Improve signatures for `maybe()`, `nullable()`, and `optional()` functions.
Add narrow signatures for `maybe()`, `nullable()`, and `optional()` unit functions.

Previously each function would instantiate `Maybe<T>`, `Nullable<T>`, or `Optional<T>`, even if the value passed to the function is always present (or always absent).

Now when a function is given an always present value it will return `Just<T>`, `Only<T>`, or `Some<T>`. And for always absent values it will return `Nothing<T>`, `Nil<T>`, or `None<T>`.

This behavior has already been the case for the methods of these classes, so this change only aligns the behavior of the unit functions.